report/requirements.txt
pandas
openpyxl
xlsxwriter
pdfplumber
google-api-python-client
google-auth-httplib2
google-auth-oauthlibrequests